Fix clang build

+++ db/memtablerep_bench.cc
@@ -184,8 +184,9 @@ class KeyGenerator {
       for (uint64_t i = 0; i < num_; ++i) {
         values_[i] = i;
       }
-      std::shuffle(values_.begin(), values_.end(),
-                   std::default_random_engine(FLAGS_seed));
+      std::shuffle(
+          values_.begin(), values_.end(),
+          std::default_random_engine(static_cast<unsigned int>(FLAGS_seed)));
     }
   }
